+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added the Forest and Flower Forest to the biome tag `#irregular_implements:nature_core_oak`
 - The Notification Interface now defaults to having no text in the title and description
 - The Player Interface no longer renders the owner's skull above itself if there's a block there
+- Glowing Mushrooms can now be used in _most_ brewing recipes that would use Glowstone (#29)
 
 ### Fixed